Addresses comments
Swaps out .filter(...)[0] with the much cleaner .find()
Rename the UI deletion action
Moves makeB64Digest to app/js/util
Cleans up parentheses
Reverts artifact deletion to original single line .filter()

Revert to Flow

Remove all unnecessary propTypes
